# Stricten getBoolean and getNonNegative