I played with Tikhons benchmarks and found that they could easily be sped up.

Before:

```
decode/array/bools                       time                 24.02 ms
decode/array/ints                        time                 82.41 ms
decode/array/longs                       time                 130.7 ms
decode/array/records                     time                 311.5 ms
```

After:

```
decode/array/bools                       time                 18.46 ms
decode/array/ints                        time                 27.34 ms
decode/array/longs                       time                 34.18 ms
decode/array/records                     time                 154.0 ms
```