@TikhonJelvis 2019-03-02
Sweet. It's great to have a large speedup with such a small change.